requirements.txt
whois
dnspython